# api_simplon_formation/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SIMPLON FORMATION",
        "description": "une api qui permet de gerer le sourcing des formation simplon",
        "version": "1.0"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "consumes": [
        "application/json"
    ],
    "paths": {
        "/api/toutesformations": {
            "get": {
                "summary": "voir toutes les formation",
                "description": "",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "examples": {
                            "application/json": ""
                        }
                    }
                },
                "tags": [
                    "Api Simplon formation"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    },
                    {
                        "in": "header",
                        "name": "",
                        "type": "string"
                    }
                ]
            }
        },
        "/api/role": {
            "get": {
                "summary": "liste des roles",
                "description": "",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "examples": {
                            "application/json": ""
                        }
                    }
                },
                "tags": [
                    "requette role"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    },
                    {
                        "in": "header",
                        "name": "Authorization",
                        "type": "string"
                    }
                ]
            },
            "post": {
                "summary": "créer un role",
                "description": "",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "examples": {
                            "application/json": ""
                        }
                    }
                },
                "tags": [
                    "requette role"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    },
                    {
                        "in": "header",
                        "name": "Authorization",
                        "type": "string"
                    }
                ]
            }
        },
...
        "/api/Candidature/refuser": {
            "get": {
                "summary": "liste candidature refusées",
                "description": "",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "examples": {
                            "application/json": ""
                        }
                    }
                },
                "tags": [
                    "Admin"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    },
                    {
                        "in": "header",
                        "name": "Authorization",
                        "type": "string"
                    }
                ]
            }
        },
        "/api/Candidature/accepter": {
            "get": {
                "summary": "liste des candidature acceptées",
                "description": "",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "examples": {
                            "application/json": ""
                        }
                    }
                },
                "tags": [
                    "Admin"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    },
                    {
                        "in": "header",
                        "name": "Authorization",
                        "type": "string"
                    }
                ]
            }
        },
...
    }
}